Scan for exit spots
These exit spots are used for the points where you start the level or where you exit transporter events.  The scanning makes a quick register of them so during the game things will go a lot faster. As a matter of fact teleportation events should go in an instant. A big of the woes of a kind of game that could not use Kthura, I suppose.